requirements.txt
# Data Generator
-e git+https://github.com/WGierke/agots#egg=agots

# Requirements for donut (https://github.com/korepwx/donut)
git+https://github.com/thu-ml/zhusuan.git
git+https://github.com/korepwx/tfsnippet.git
git+https://github.com/korepwx/donut.git

# Requirements for dagmm (https://github.com/danieltan07/dagmm)
torch>=0.4.0
torchvision
h5py

# Remaining requirements that should be installed if not already done so
numpy
pandas
tqdm
scikit-learn>=0.19.1
flake8
scipy
matplotlib
progressbar2
